// front/src/types/store/dashboard.ts
export interface StoreChats {
  totalInstallments: number;
  activeInstallments: number;
  completedInstallments: number;
  overdueInstallments: number;
  newThisMonth: number;
  monthlyRevenue: number;
  revenueGrowth: number;
  overduePayments: number;
  totalActiveAmount: number;
  totalCustomers: number;
  blacklistedCustomers: number;
  overdueCustomers: number;
  activeCustomers: number;
}